Fix video not saving issue
